\name{boxplot-methods}
\docType{methods}
\alias{boxplot-methods}
\alias{boxplot,EList-method}
\alias{boxplot,RGList-method}
\title{Methods for Function boxplot}
\description{
  Methods for function \code{boxplot} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    boxplot for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      boxplot for \code{RGList} object
}
}}
\keyword{methods}
\keyword{graphics}
