\name{exprs-methods}
\docType{methods}
\alias{exprs-methods}
\alias{exprs,EList-method}
\alias{exprs,RGList-method}
\title{Methods for Function exprs}
\description{
  Methods for function \code{exprs} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    exprs for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      exprs for \code{RGList} object
}
}}
\keyword{methods}
