\name{featureNames-methods}
\docType{methods}
\alias{featureNames-methods}
\alias{featureNames,EList-method}
\alias{featureNames,RGList-method}
\title{Methods for Function featureNames}
\description{
  Methods for function \code{featureNames} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    featureNames for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      featureNames for \code{RGList} object
}
}}
\keyword{methods}
