\name{preproc<--methods}
\docType{methods}
\alias{preproc<--methods}
\alias{preproc<-,EList-method}
\alias{preproc<-,RGList-method}
\title{Methods for Function preproc}
\description{
  Methods for function \code{preproc<-} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    preproc for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      preproc for \code{RGList} object
}
}}
\keyword{methods}

