\name{probeNames-methods}
\docType{methods}
\alias{probeNames-methods}
\alias{probeNames,EList-method}
\alias{probeNames,RGList-method}
\title{Methods for Function probeNames}
\description{
  Methods for function \code{probeNames} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    probeNames for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      probeNames for \code{RGList} object
}
}}
\keyword{methods}
