\name{sampleNames-methods}
\docType{methods}
\alias{sampleNames-methods}
\alias{sampleNames,EList-method}
\alias{sampleNames,RGList-method}
\title{Methods for Function sampleNames}
\description{
  Methods for function \code{sampleNames} for objects
  of class  \code{EList} and \code{RGList} 
}
\section{Methods}{
\describe{

  \item{\code{signature(x = "EList")}}{
    sampleNames for \code{EList} object
    
}

\item{\code{signature(x = "RGList")}}{
      sampleNames for \code{RGList} object
}
}}
\keyword{methods}
